full python game is running

+++ engine/src/main/battlecode/crossplay/CrossPlay.java
@@ -15,13 +15,18 @@ import battlecode.common.*;
  */
 public class CrossPlay {
     public static final String
-        CROSS_PLAY_DIR = "crossplay", // temporary directory for cross-play files
+        CROSS_PLAY_DIR = "crossplay_temp", // temporary directory for cross-play files
         MESSAGE_FILE_JAVA = "messages_java.json", // messages from the java engine
         MESSAGE_FILE_OTHER = "messages_other.json", // messages from the other language's runner script
         LOCK_FILE_JAVA = "lock_java.txt", // lock file created by the java engine
         LOCK_FILE_OTHER = "lock_other.txt"; // lock file created by the other language's runner script
 
-    ArrayList<Object> objects;
+    private ArrayList<Object> objects;
+    private RobotController rc;
+    private int roundNum;
+    private Team team;
+    private int id;
+    private CrossPlayReference rcRef;
 
     public CrossPlay() {
         this.objects = new ArrayList<>();
@@ -31,7 +36,7 @@ public class CrossPlay {
         this.objects.clear();
     }
 
-    private void resetFiles() {
+    public static void resetFiles() {
         try {
             Path crossPlayDir = Paths.get(CROSS_PLAY_DIR);
 
@@ -48,7 +53,7 @@ public class CrossPlay {
         }
     }
 
-    private void clearTempFiles() {
+    public static void clearTempFiles() {
         try {
             Path crossPlayDir = Paths.get(CROSS_PLAY_DIR);
 
@@ -111,20 +116,19 @@ public class CrossPlay {
         return ref;
     }
 
-    public void run() {
-        clearObjects();
-        resetFiles();
-
+    public void runMessagePassing() {
         Path crossPlayDir = Paths.get(CROSS_PLAY_DIR);
         Path javaMessagePath = crossPlayDir.resolve(MESSAGE_FILE_JAVA);
         Path otherMessagePath = crossPlayDir.resolve(MESSAGE_FILE_OTHER);
         Path javaLockPath = crossPlayDir.resolve(LOCK_FILE_JAVA);
         Path otherLockPath = crossPlayDir.resolve(LOCK_FILE_OTHER);
+        // System.out.println("Waiting for message Python -> Java...");
 
         while (true) {
             try {
                 if (!Files.exists(otherMessagePath) || Files.exists(javaMessagePath) || Files.exists(otherLockPath)) {
-                    Thread.sleep(0, 100000); // sleep for 0.1 ms
+                    Thread.sleep(100, 0); // TODO make shorter: sleep for 0.1 s
+                    // System.out.println("Still waiting for message Python -> Java...");
                     continue;
                 }
 
@@ -133,31 +137,35 @@ public class CrossPlay {
                         + " This should never happen under normal operation.");
                 }
 
-                // create java lock file
                 Files.createFile(javaLockPath);
-                // read other language's messages
                 String messageContent = Files.readString(otherMessagePath);
                 JSONObject messageJson = new JSONObject(messageContent);
                 CrossPlayMessage message = CrossPlayMessage.fromJson(messageJson);
-                CrossPlayObject result = processMessage(message);
-                // delete other language's message file
-                Files.delete(otherMessagePath);
 
-                // write to java message file
+                // System.out.println("Received message Python -> Java: " + messageJson.toString());
+
+                if (message.method == CrossPlayMethod.TERMINATE) {
+                    Files.delete(otherMessagePath);
+                    Files.delete(javaLockPath);
+                    // System.out.println("Received terminate message, ending cross-play message passing.");
+                    break;
+                }
+
+                CrossPlayObject result = processMessage(message);
                 String resultContent = result.toJson().toString();
                 Files.writeString(javaMessagePath, resultContent);
 
-                // delete java lock file
+                Files.delete(otherMessagePath);
                 Files.delete(javaLockPath);
-                break;
+
+                // System.out.println("Sent response Java -> Python: " + resultContent);
+                // System.out.println("Waiting for message Python -> Java...");
             } catch (InterruptedException e) {
                 throw new CrossPlayException("Cross-play message passing thread was interrupted.");
             } catch (IOException e) {
                 throw new CrossPlayException("Failed to read other language's cross-play message file.");
             }
         }
-
-        clearTempFiles();
     }
 
     // private JSONArray processJsonMessages(JSONArray json) {
@@ -199,6 +207,17 @@ public class CrossPlay {
         switch (message.method) {
             case INVALID:
                 throw new CrossPlayException("Received invalid cross-play method!");
+            case TERMINATE:
+                throw new CrossPlayException("Terminate messages should be handled outside of processMessage.");
+            case START_TURN:
+                result = new CrossPlayLiteral(CrossPlayObjectType.ARRAY, new CrossPlayObject[] {
+                    this.rcRef,
+                    new CrossPlayLiteral(CrossPlayObjectType.INTEGER, this.roundNum),
+                    new CrossPlayLiteral(CrossPlayObjectType.TEAM, this.team),
+                    new CrossPlayLiteral(CrossPlayObjectType.INTEGER, this.id),
+                    new CrossPlayLiteral(CrossPlayObjectType.BOOLEAN, false),
+                });
+                break;
             case RC_GET_ROUND_NUM:
                 rc = this.<RobotController>getObject(computedParams[0]);
                 result = new CrossPlayLiteral(CrossPlayObjectType.INTEGER, rc.getRoundNum());
@@ -214,10 +233,29 @@ public class CrossPlay {
             case LOG:
                 String msg = getLiteralValue(computedParams[0]);
                 System.out.println(msg);
+                result = new CrossPlayLiteral(CrossPlayObjectType.NULL, null);
+                break;
             default:
                 throw new CrossPlayException("Received unknown cross-play method: " + message.method);
         }
 
         return result;
+    }
+
+    public void playTurn(RobotController rc) {
+        this.rc = rc;
+        this.roundNum = rc.getRoundNum();
+        this.team = rc.getTeam();
+        this.id = rc.getID();
+
+        if (this.roundNum == 1) {
+            clearObjects();
+            this.rcRef = setNextObject(CrossPlayObjectType.ROBOT_CONTROLLER, this.rc);
+            System.out.println("Cross-play bot initialized!");
+        } else {
+            this.rcRef = new CrossPlayReference(CrossPlayObjectType.ROBOT_CONTROLLER, 0);
+        }
+        
+        runMessagePassing();
     }
 }
